Update clipped value for analogRead
Set the default clipped value of A4 and A5 pins when the HPADC gain
setting is -6dB.

Add new API analogReadMap() that can change the clipped value of min
and max value. As the special option, if the min value is the same
value as the max value, analogRead() returns the raw value of signed
short type without clipping.