File: Fix issue cannot overwrite file on Flash
As the POSIX specification, to writing with O_APPEND is always appended
to the end of the file. This is not compliant with the Arduino File
specification. Therefore, this commit removes O_APPEND from FILE_WRITE
mode and adds to seek to the end of the file after file open.